+++ examples/nunjucks/components/context/context.spec.js
@@ -1,0 +1,11 @@
+const { context } = require('@frctl/adapter-tests');
+
+const fractal = require('../../fractal.config.js');
+
+describe('context', () => {
+    beforeEach(async () => {
+        await fractal.load();
+    });
+
+    context(fractal);
+});